fix(build): usize 타입 캐스팅 오류 수정
- lib.rs에서 format_number 호출 시 usize를 u64로 명시적 변환

+++ src-tauri/src/lib.rs
@@ -192,8 +192,8 @@ async fn sync_dry_run(
         Ok(result) => {
             let msg = format!(
                 "Dry run completed.\nTo copy: {} files\nTo delete: {} files\nTotal size: {}",
-                format_number(result.files_to_copy), 
-                format_number(result.files_to_delete), 
+                format_number(result.files_to_copy as u64), 
+                format_number(result.files_to_delete as u64), 
                 format_bytes(result.bytes_to_copy)
             );
             state.log_manager.log("success", &msg, Some(task_id));
